# feat: Enhance bulk insert speed using copy command