I've been doing very large loads into PostgreSQL, and found `target-postgres` to be slow at bulk inserts. In PostgreSQL, the [COPY command](https://www.postgresql.org/docs/current/sql-copy.html) is preferable for high performance loading of data.

This PR changes the implementation of `bulk_insert_records()` by using `COPY` with CSV data instead of an `INSERT` statement with values. When combined with larger batches (`batch_size_rows`), I am able to achieve 2 to 10× bulk insert speeds compared to the existing implementation.

Here's an example difference for moving 1 million rows via `tap-snowflake target-postgres`.

```
new  48.46s user 4.19s system 19% cpu  4:36.91 total
old 189.69s user 6.76s system 28% cpu 11:33.89 total
```

**Useful Links:**

* [14.4.2 Populating a Database — Use COPY](https://www.postgresql.org/docs/current/populate.html#POPULATE-COPY-FROM)
* [How does COPY work and why is it so much faster than INSERT?](https://stackoverflow.com/questions/46715354/how-does-copy-work-and-why-is-it-so-much-faster-than-insert)