@visch 2024-06-18
> @visch did we document somewhere the steps needed to regenerate the SSL cert files so CI can pass?



> Member

We didn't :/ we should have at least linked to https://www.postgresql.org/docs/current/ssl-tcp.html#SSL-CERTIFICATE-CREATION which has the steps. Really our steps should setup them to be 250 years or something.